finished ticket 15 - delete a tag
